gpuav: Try to fix crash
